+++ Views/Home/Checkout.cshtml
@@ -3,6 +3,7 @@
 }
 <script src="https://www.paypal.com/sdk/js?client-id=AQJ-nomYmdcrcJJa_kZYrA-Lzy1AdtFLCMYgwwyBhOxumCXDaQ4r7VtkGAcyDMqZhfVvz27iS7mmbgXW&currency=NZD"></script>
 <script data-sdk-integration-source="integrationbuilder_ac"></script>
+<script src="~/js/paypal.js"></script>
 <div id="paypal-button-container"></div>
 <script>
     const fundingSources = [
